# plataforma-recursos/services/interface/swagger.yaml
openapi: 3.0.3
info:
  title: EW2026 Interface Service
  version: 1.0.0
  description: Interface web server-side (Pug) da plataforma de recursos educativos.
servers:
  - url: http://localhost:16020
    description: Interface via gateway
  - url: http://localhost:16026
    description: Interface direta
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ew.sid
  schemas:
...
    ErrorPage:
      type: object
      properties:
        title:
          type: string
          example: Erro interno
        message:
          type: string
          example: Ocorreu um erro inesperado. Tente novamente.
paths:
...
  /resources:
    get:
      tags: [Resources]
      summary: Listar recursos com filtros e paginação
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: tipo
          in: query
          schema:
            type: string
        - name: ano
          in: query
          schema:
            type: string
        - name: tema
          in: query
          schema:
            type: string
        - name: hashtag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTML de listagem de recursos
          content:
            text/html:
              schema:
                type: string
  /resources/new:
    get:
      tags: [Resources]
      summary: Formulário de submissão de recurso
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML do formulário
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirecionamento caso não autenticado/sem permissões
    post:
      tags: [Resources]
      summary: Submeter SIP (ZIP) para ingestão
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [sip]
              properties:
                sip:
                  type: string
                  format: binary
      responses:
        '302':
          description: Redireciona para detalhe do recurso ou para o formulário em erro
  /resources/{id}:
    get:
      tags: [Resources]
      summary: Ver detalhe de recurso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTML com detalhe do recurso
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Página de erro quando recurso não existe
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/ErrorPage'
  /resources/{id}/dip:
    get:
      tags: [Resources]
      summary: Download do DIP do recurso
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ficheiro ZIP do DIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Sem autenticação para recurso privado
        '403':
          description: Sem permissões
        '404':
          description: Recurso não encontrado
  /resources/{id}/ratings:
    post:
      tags: [Resources]
      summary: Submeter avaliação de recurso
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [stars]
              properties:
                stars:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
      responses:
        '302':
          description: Redireciona para detalhe do recurso
...